+++ frappe_manager/commands/self/__init__.py
@@ -4,6 +4,7 @@ import typer
 from typer_examples import install
 
 from frappe_manager.commands.self.stop import stop
+from frappe_manager.commands.self.uninstall import uninstall
 from frappe_manager.commands.self.update_images import update_images
 from frappe_manager.commands.self.upgrade import upgrade
 
@@ -14,3 +15,4 @@ install(self_app)
 self_app.command()(upgrade)
 self_app.command(name="update-images")(update_images)
 self_app.command()(stop)
+self_app.command()(uninstall)